+++ test/i18n/menus.test.ts
@@ -204,6 +204,7 @@ const DATA = [
 	'Issue',
 	'Idea',
 	'Milestone',
+	'Resource',
 	'Deliverable',
 	'New',
 	'Active',